Bug #636026 <update lsc and cmc new interpolation alg>
[root cause  ] new lsc and cmc10 interpolation alg for sharkl2
[changes     ] add new interpolation alg
[side effects] NA
[self test   ]<board, Test cases summary>:sharkl2,preview/capture OK
[reviewers   ] eddy.wei,xiang.bao,chunji.che,feng.tian
[change_type ] feature_bugfix
[tag_product ] specific

	modified:   isp2.1/common/inc/isp_com.h
	modified:   isp2.1/param_manager/isp_com_alg.c
	modified:   isp2.1/param_manager/src/isp_blk_2d_lsc.c
	modified:   isp2.1/param_manager/src/isp_blk_cmc10.c

+++ isp2.1/param_manager/isp_com_alg.c
@@ -391,7 +391,7 @@ static void _interp_int14(uint16_t *dst, uint16_t *src[2], uint16_t weight[2], u
 	uint32_t data_bytes = 0;
 	uint32_t i = 0;
 
-	data_bytes = data_num * sizeof(uint32_t);
+	data_bytes = data_num * sizeof(uint16_t);
 
 	if (INTERP_WEIGHT_UNIT == weight[0]) {
 		memcpy(dst, src[0], data_bytes);
